requirements/requirements.txt
python-dotenv
flask
Flask-PyMongo
sentry-sdk[flask]
jsonschema
ndjson
